Changes
Have mocked up the first page for click through, if we can get this right we can just replicate and drop in images for all 12.

Can't have the snow and get buttons to work at the same time, don't know if it's creating a layer on top?

When I added links to the whole door element it makes the door move? We'd need it to be on the whole door for it to work for mobile users.

Please feel free to amend anything I've done to make it work better!